refactor: 패키지 구조 수정
비즈니스 로직을 담당하는 model과 사용자와의 상호작용을 담당하는 view로 구조 수정
Util 클래스를 Utilization으로 이름 변경
