+++ src/main/java/baseball/model/util/Constant.java
@@ -1,4 +1,4 @@
-package baseball.util;
+package baseball.model.util;
 
 public interface Constant {
     String GAME_START_MESSAGE = "숫자 야구 게임을 시작합니다.";
