+++ src/main/java/baseball/model/util/Utilization.java
@@ -1,19 +1,18 @@
-package baseball.util;
+package baseball.model.util;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
 
-import static baseball.util.Constant.*;
 import static camp.nextstep.edu.missionutils.Randoms.pickNumberInRange;
 
-public abstract class Util {
+public abstract class Utilization {
     public static List<Integer> pickRandomNumberList() {
         List<Integer> randomNumbers = new ArrayList<>();
 
-        while (randomNumbers.size() < NUMBER_LENGTH) {
-            int randomNumber = pickNumberInRange(MIN_NUMBER_VALUE, MAX_NUMBER_VALUE);
+        while (randomNumbers.size() < Constant.NUMBER_LENGTH) {
+            int randomNumber = pickNumberInRange(Constant.MIN_NUMBER_VALUE, Constant.MAX_NUMBER_VALUE);
             if (!randomNumbers.contains(randomNumber)) {
                 randomNumbers.add(randomNumber);
             }
