+++ src/main/java/baseball/view/Game.java
@@ -1,0 +1,4 @@
+package baseball.view;
+
+public class Game {
+}
